requirements.txt
pyyaml>=5.3.1
scipy>=1.5.0
tensorboard>=2.2.1t0
torchmeta==1.8.0
trimesh
plyfile
matplotlib
configargparse
scikit-image
mesh-to-sdf
pyrender
tqdm
h5py